test: improve EL-related tests
Scope:
- Added unit tests for ELs in the context of sequential/parallel multi-instance activities
- Improve tests structure:
 - Move EL-related tests to `listener.execution` package
 - Move task related tests from `ExecutionListenerTest` to `ExecutionListenerTaskElementsTest` class
- Improved existing tests for Execution Listeners, cover several missing scenarios.
 - Add scenario to check that subsequent element can access variables produced by `end` EL job
